Autotest/cicd/python/requirements.txt
# =============================================================================
# Dependencias Python para CI/CD Pipeline
# =============================================================================
# Compatible con Python 3.6+
# Instalar con: pip3.6 install --user -r requirements.txt
#
# NOTA: No usamos pyvmomi porque requiere Python 3.9+
#       En su lugar, usamos la API REST de vCenter directamente con requests

# HTTP requests para APIs REST (vCenter y SonarQube)
requests>=2.20.0,<3.0.0

# Parser YAML para configuración
PyYAML>=5.0,<6.1

# Para manejo de certificados SSL
certifi>=2020.0.0

# Codificación de caracteres
charset-normalizer>=2.0.0,<3.0.0

# URLs internacionalizadas
idna>=2.5,<4.0

# HTTP library
urllib3>=1.21.1,<2.0.0
